reorganized into tabs

+++ data_processing.py
@@ -1,23 +1,25 @@
 import yfinance as yf
 from datetime import datetime
 import pandas as pd
-
-START_DATE = "2018-01-01"
-STOCK_LIST =  ['NVDA', 'TSLA', 'AMZN', 'MSFT', 'COST', 'GME', 'PFE', 'LLY']
+from constants import STOCK_DATA_DICT, START_DATE, TODAY
+import os
 
 def fetch_data():
-    today = datetime.today()
-
-    for stock_symbol in STOCK_LIST:
-        stock_data = yf.download(stock_symbol, start=START_DATE, end=today)    
+    # Check if the 'data' directory exists, raise an error if it doesn't
+    if not os.path.exists('data'):
+        raise FileNotFoundError("The 'data' directory does not exist. Please create it before running the script.")
+    
+    for company, stock_symbol in STOCK_DATA_DICT.items():
+        stock_data = yf.download(stock_symbol, start=START_DATE, end=TODAY)
+        stock_data.to_csv(f"data/{stock_symbol}.csv")
         stock_data.to_csv(f"data/{stock_symbol.lower()}_stock.csv")
 
 def _count_null(df):
     return df.isnull().sum().sum()
 
 def check_null():
-    for index in STOCK_LIST:
-        index_df = pd.read_csv(f"data/{index.lower()}_stock.csv")
+    for company, index in STOCK_DATA_DICT.items():
+        index_df = pd.read_csv(f"data/{index}.csv")
         total_nulls = _count_null(index_df)
         if total_nulls > 0:
             print(f"{index} has {total_nulls} null entries")
